backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
aiosqlite>=0.20.0,<0.21.0
redis[hiredis]>=5.0.0,<6.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9
siwe>=4.0.0,<5.0.0
web3>=7.0.0,<8.0.0
eth-account>=0.13.0,<0.14.0
fal-client>=0.5.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.28.0,<1.0.0
